+++ src/frontend/src/pages/Index/Profile/UserPanel.tsx
@@ -25,9 +25,9 @@ import { useState } from 'react';
 import { api, queryClient } from '../../../App';
 import { EditButton } from '../../../components/items/EditButton';
 import { PlaceholderPill } from '../../../components/items/Placeholder';
-import { useLocalState } from '../../../context/LocalState';
 import { SizeMarks } from '../../../defaults';
 import { InvenTreeStyle } from '../../../globalStyle';
+import { useLocalState } from '../../../states/LocalState';
 
 export function UserPanel() {
   // view
