+++ src/frontend/src/pages/NotFound.tsx
@@ -3,7 +3,7 @@ import { Button, Center, Container, Stack, Text, Title } from '@mantine/core';
 import { IconArrowBack } from '@tabler/icons-react';
 import { useNavigate } from 'react-router-dom';
 
-import { LanguageContext } from '../context/LanguageContext';
+import { LanguageContext } from '../contexts/LanguageContext';
 
 export default function NotFound() {
   const navigate = useNavigate();
